Refactor: Add authentication and authorization middleware to products router

+++ src/components/products/v1/web/products.routes.ts
@@ -1,33 +1,61 @@
-import express from 'express';
-import { ProductsController } from './products.controllers';
-import { ProductService } from '../services/products.service';
+import express from "express";
+import { ProductsController } from "./products.controllers";
+import { ProductService } from "../services/products.service";
+import { Permission } from "../../../../config/permissions";
+import { authorize } from "../../../auth/v1/middlewares/authorize";
+import { authenticate } from "../../../auth/v1/middlewares/authenticate";
 
 const router = express.Router();
 const productService = new ProductService();
 const productsController = new ProductsController(productService);
 
-router.get('/', async (req, res) => {
-  if (req.query.query) {
-    await productsController.searchProducts(req, res);
-  } else {
-    await productsController.getProducts(req, res);
+router.get(
+  "/",
+  authenticate,
+  authorize(Permission.VIEW_PRODUCT),
+  async (req, res) => {
+    if (req.query.query) {
+      await productsController.searchProducts(req, res);
+    } else {
+      await productsController.getProducts(req, res);
+    }
   }
-});
+);
 
-router.get('/:id', async (req, res) => {
-  await productsController.getProductById(req, res);
-});
+router.get(
+  "/:id",
+  authenticate,
+  authorize(Permission.VIEW_PRODUCT),
+  async (req, res) => {
+    await productsController.getProductById(req, res);
+  }
+);
 
-router.post('/', async (req, res) => {
-  await productsController.createProduct(req, res);
-});
+router.post(
+  "/",
+  authenticate,
+  authorize(Permission.CREATE_PRODUCT),
+  async (req, res) => {
+    await productsController.createProduct(req, res);
+  }
+);
 
-router.put('/:id', async (req, res) => {
-  await productsController.updateProductById(req, res);
-});
+router.put(
+  "/:id",
+  authenticate,
+  authorize(Permission.UPDATE_PRODUCT),
+  async (req, res) => {
+    await productsController.updateProductById(req, res);
+  }
+);
 
-router.delete('/:id', async (req, res) => {
-  await productsController.deleteProductById(req, res);
-});
+router.delete(
+  "/:id",
+  authenticate,
+  authorize(Permission.DELETE_PRODUCT),
+  async (req, res) => {
+    await productsController.deleteProductById(req, res);
+  }
+);
 
 export default router;
